Added averaging for latitude and longitude and clamped avgspeed calculation
Added averaging to gpspolling, it takes n+1 points during the interval of gpsPollRate and averages the lat and long parameters to give a more stable result. n = gpsAveragePointsInInterval. Additionally moved the avg speed calculation to happen only when the distance is updated to prevent the avgSpeed from oscillating with gps.